100 bug button text and background z fighting due to missing layer offsets (#102)
* Fix: Remove bug with icon and text position

* Feat: Add another button

* Feat: Add some fix to text rendering

+++ libraries/evan/sources/glfw/DesktopBackend.cpp
@@ -40,7 +40,8 @@ static VKAPI_ATTR VkBool32 VKAPI_CALL defaultDebugCallback(
 	static std::unique_ptr<utility::logging::Logger> logger = nullptr;
 	std::string_view message;
 	if (!logger) {
-		logger = std::make_unique<utility::logging::DefaultLogger>("VulkanValidation");
+		logger = std::make_unique<utility::logging::DefaultLogger>(
+			"VulkanValidation");
 	}
 
 	switch (messageType) {
@@ -719,8 +720,8 @@ void evan::DesktopBackend::setupCallbackEvent(const IPlatform &platform)
 
 			auto event = std::make_shared<utility::event::MouseMotionEvent>();
 			event->setPosition(utility::event::MouseMotionEvent::MousePosition {
-				static_cast<unsigned int>(xpos),
-				static_cast<unsigned int>(ypos) });
+				static_cast<float>(xpos),
+				static_cast<float>(ypos) });
 			self->_mouseMotionEvents.push_back(std::move(event));
 		});
